Move `home` function definition to `routes/index`
Use the more descriptive `routes.rs` rather than `routes/mod.rs`
